Fixed document symbol slowness whith long function
When trying to match some function, the document symbol provider could
completely hang because the regex was backtracking, and it took too long
to complete.

I've simplified the regex. It could give some false positive with some
incorrectly written functions, but I don't think this is a big issue.

Fixes #25