Improved block comments behavior
Before, adding block comments was not adding the comments correctly. It
was adding block comments that were not in a correct format for Matlab.

As proposed by @CelsoReyes in #31, I've added line breaks before and
after the %{ and %} characters. This makes the block comments be
recognized correctly by Matlab, but has the unfortunate drawback that we
cannot remove them correctly.

Unfortunately that's all we can do for now.